chore: re exports and react cleanup
