+++ packages/nextjs/tsconfig.json
@@ -28,6 +28,7 @@
     "site.d.ts",
     "search.d.ts",
     "client.d.ts",
+    "atoms.d.ts",
     "utils.d.ts",
     "tools.d.ts",
     "config-cli.d.ts",
